+++ packages/wallet-sdk/src/sign/walletlink/relay/WalletLinkRelay.test.ts
@@ -3,6 +3,7 @@
 import { WalletLinkCipher } from './connection/WalletLinkCipher';
 import { WalletLinkConnection } from './connection/WalletLinkConnection';
 import { WalletLinkWebSocket } from './connection/WalletLinkWebSocket';
+import { WALLET_USER_NAME_KEY } from './constants';
 import { ServerMessage } from './type/ServerMessage';
 import { WalletLinkSessionConfig } from './type/WalletLinkSessionConfig';
 import { WalletLinkRelay } from './WalletLinkRelay';
@@ -100,7 +101,7 @@ describe('WalletLinkRelay', () => {
       });
 
       expect(metadataUpdatedSpy).toHaveBeenCalledWith(
-        'walletUsername',
+        WALLET_USER_NAME_KEY,
         await decryptMock(sessionConfig.metadata.WalletUsername)
       );
     });
