+++ packages/wallet-sdk/src/sign/walletlink/relay/constants.ts
@@ -1,0 +1,3 @@
+export const WALLET_USER_NAME_KEY = 'walletUsername';
+export const LOCAL_STORAGE_ADDRESSES_KEY = 'Addresses';
+export const APP_VERSION_KEY = 'AppVersion';
